ipa: libipa: lsc: Quantize lsc gains in IPA
The libIPA Lsc algorithm currently assumes the register format of
the RkISP1 platform (16 bits Q2.10 format), as that's where the
implementation has been derived from.

Move the register quantization to the plaltform IPA and use 'floats'
as the exchange format for the table-based and polynomial-based
LSC implementations.

+++ src/ipa/libipa/lsc_base.h
@@ -26,7 +26,7 @@ namespace ipa {
 
 namespace lsc {
 
-using Components = std::map<std::string, std::vector<uint16_t>>;
+using Components = std::map<std::string, std::vector<float>>;
 using ComponentsMap = std::map<unsigned int, Components>;
 
 } /* namespace lsc */
